Rename "parse_input" -> "process_input" and "check_output" -> "process_output"
