+++ core_backend/app/config.py
@@ -55,7 +55,7 @@ ALIGN_SCORE_THRESHOLD = os.environ.get("ALIGN_SCORE_THRESHOLD", 0.7)
 # Method: LLM, AlignScore, or None
 ALIGN_SCORE_METHOD = os.environ.get("ALIGN_SCORE_METHOD", "LLM")
 # if AlignScore, set ALIGN_SCORE_API. If LLM, set LITELLM_MODEL_ALIGNSCORE above.
-ALIGN_SCORE_API = os.environ.get("ALIGN_SCORE_API")
+ALIGN_SCORE_API = os.environ.get("ALIGN_SCORE_API", "")
 
 # Backend paths
 BACKEND_ROOT_PATH = os.environ.get("BACKEND_ROOT_PATH", "")
