+++ core_backend/app/question_answer/routers.py
@@ -8,16 +8,16 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from ..auth.dependencies import auth_bearer_token
 from ..contents.models import get_similar_content_async, update_votes_in_db
 from ..database import get_async_session
-from ..llm_call.check_output import check_align_score__after
 from ..llm_call.llm_prompts import ANSWER_FAILURE_MESSAGE
 from ..llm_call.llm_rag import get_llm_rag_answer
-from ..llm_call.parse_input import (
+from ..llm_call.process_input import (
     classify_on_off_topic__before,
     classify_safety__before,
     identify_language__before,
     paraphrase_question__before,
     translate_question__before,
 )
+from ..llm_call.process_output import check_align_score__after
 from ..utils import generate_secret_key
 from .config import N_TOP_CONTENT_FOR_RAG, N_TOP_CONTENT_FOR_SEARCH
 from .models import (
